+++ amazon/s3/client/src/main/kotlin/org/http4k/connect/amazon/s3/action/S3BucketAction.kt
@@ -1,5 +1,7 @@
 package org.http4k.connect.amazon.s3.action
 
+import dev.forkhandles.result4k.Result
 import org.http4k.connect.Action
+import org.http4k.connect.RemoteFailure
 
-interface S3BucketAction<R> : Action<R>
+interface S3BucketAction<R> : Action<Result<R, RemoteFailure>>