+++ amazon/sqs/client/src/main/kotlin/org/http4k/connect/amazon/sqs/action/SQSAction.kt
@@ -1,6 +1,8 @@
 package org.http4k.connect.amazon.sqs.action
 
+import dev.forkhandles.result4k.Result
 import org.http4k.connect.Action
+import org.http4k.connect.RemoteFailure
 import org.http4k.core.ContentType.Companion.APPLICATION_FORM_URLENCODED
 import org.http4k.core.Method.POST
 import org.http4k.core.Request
@@ -10,7 +12,7 @@ import org.http4k.core.with
 import org.http4k.lens.Header.CONTENT_TYPE
 
 abstract class SQSAction<R>(private val action: String, private vararg val mappings: Pair<String, String>?) :
-    Action<R> {
+    Action<Result<R, RemoteFailure>> {
     override fun toRequest() =
         (listOf("Action" to action, "Version" to "2012-11-05") + mappings)
             .filterNotNull()