+++ amazon/sts/client/src/main/kotlin/org/http4k/connect/amazon/sts/action/STSAction.kt
@@ -1,5 +1,7 @@
 package org.http4k.connect.amazon.sts.action
 
+import dev.forkhandles.result4k.Result
 import org.http4k.connect.Action
+import org.http4k.connect.RemoteFailure
 
-interface STSAction<R> : Action<R>
+interface STSAction<R> : Action<Result<R, RemoteFailure>>