+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
+### v2.14.0.0
+- **http4k-connect-*** : [Breaking] Changed Result type on Action to be generic to support other programming models. This will only affect users who are implementing their own adapters. To fix, change: 
+```kotlin
+interface MyAdapter<R> : Action<R>
+// to 
+interface MyAdapter<R> : Action<Result<R, RemoteFailure>>
+```
+
 ### v2.13.0.1
 - **http4k-connect-amazon-s3-fake* : Send response XML as well as status code on errors.